requirements.txt
mysql-connector-python==8.0.33
flask==3.0.0
pymongo==4.8.0
flask-cors==5.0.0
